Make a BookmarkSearch kwargs mixin
